+++ src/midi/index.ts
@@ -2,4 +2,5 @@ export * from "./cc-rate-limit";
 export * from "./connection";
 export * from "./errors";
 export * from "./ports";
+export * from "./reassembly";
 export * from "./request-response";